ImgToVideo&Inv_App/requirements.txt
numpy
streamlit
opencv-python-headless
PyDrive
azure-storage-blob